+++ tests/test_CiscoIosInterface.py
@@ -4,6 +4,9 @@ import json
 from ccutils.ccparser import ConfigParser
 from ccutils.utils.common_utils import jprint
 
+DEBUG = False
+VERBOSITY = 5 if DEBUG else 3
+
 
 class TestL3Interface(unittest.TestCase):
     test_file_base = "cisco_ios_interface_l3_tests"
@@ -13,6 +16,15 @@ class TestL3Interface(unittest.TestCase):
     config.minimal_results = True
     results = json.loads(result_file_path.read_text())
 
+    def test_ipv4_addresses(self):
+        tested_interfaces = ["Vlan100", "Vlan101", "Vlan102"]
+
+        for interface in tested_interfaces:
+            with self.subTest(msg=interface):
+                interface_line = [x for x in self.config.interface_lines if x.name == interface][0]
+                want = self.results["interfaces"][interface]["ipv4"]["addresses"]
+                have = interface_line.ipv4_addresses
+                self.assertEqual(want, have)
 
     def test_isis(self):
         tested_interfaces = ["TenGigabitEthernet0/0/8"]
@@ -20,7 +32,6 @@ class TestL3Interface(unittest.TestCase):
         for interface in tested_interfaces:
             with self.subTest(msg=interface):
                 interface_line = [x for x in self.config.interface_lines if x.name == interface][0]
-                print(interface_line)
                 want = self.results["interfaces"][interface]["isis"]
                 have = interface_line.isis
                 self.assertEqual(want, have)
@@ -31,9 +42,18 @@ class TestL3Interface(unittest.TestCase):
         for interface in tested_interfaces:
             with self.subTest(msg=interface):
                 interface_line = [x for x in self.config.interface_lines if x.name == interface][0]
-                print(interface_line)
                 want = self.results["interfaces"][interface]["bfd"]
                 have = interface_line.bfd
+                self.assertEqual(want, have)
+
+    def test_standby(self):
+        tested_interfaces = ["Vlan100", "Vlan101", "Vlan102"]
+
+        for interface in tested_interfaces:
+            with self.subTest(msg=interface):
+                interface_line = [x for x in self.config.interface_lines if x.name == interface][0]
+                want = self.results["interfaces"][interface]["standby"]
+                have = interface_line.standby
                 self.assertEqual(want, have)
 
 
@@ -50,7 +70,6 @@ class TestL2Interface(unittest.TestCase):
         for interface in tested_interfaces:
             with self.subTest(msg=interface):
                 interface_line = [x for x in self.config.interface_lines if x.name == interface][0]
-                print(interface_line)
                 want = self.results["interfaces"][interface]["dhcp_snooping"]
                 have = interface_line.dhcp_snooping
                 self.assertEqual(want, have)