+++ tests/test_ConfigParser.py
@@ -4,6 +4,9 @@ from ccutils.ccparser import ConfigParser, CiscoIosParser
 from ccutils.utils.common_utils import jprint
 import json
 
+DEBUG = False
+VERBOSITY = 5 if DEBUG else 3
+
 
 class TestConfigParser01(unittest.TestCase):
 
@@ -17,60 +20,60 @@ class TestConfigParser01(unittest.TestCase):
         }
         for test, instance in tests.items():
             with self.subTest(msg=test):
-                parser = ConfigParser(config=self.config_path, device_type=test)
+                parser = ConfigParser(config=self.config_path, device_type=test, verbosity=VERBOSITY)
                 self.assertIsInstance(parser, instance)
 
     def test_name_servers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["name_servers"]
         have = parser.name_servers
         self.assertEqual(want, have)
 
     def test_ntp_servers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_servers"]
         have = parser.ntp_servers
         self.assertEqual(want, have)
 
     def test_ntp_peers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_peers"]
         have = parser.ntp_peers
         self.assertEqual(want, have)
 
     def test_ntp_access_groups(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_access_groups"]
         have = parser.ntp_access_groups
         self.assertEqual(want, have)
 
     def test_ntp_authentication_keys(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_authentication_keys"]
         have = parser.ntp_authentication_keys
         self.assertEqual(want, have)
 
     def test_ntp_trusted_keys(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_trusted_keys"]
         have = parser.ntp_trusted_keys
         self.assertEqual(want, have)
 
     def test_ntp_global_params(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["ntp_global"]
         have = parser.ntp_global_params
         self.assertEqual(want, have)
 
     def test_ntp(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = {
             "source": "Loopback1000",
@@ -85,7 +88,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_logging_servers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["logging_servers"]
         have = parser.logging_servers
@@ -93,7 +96,7 @@ class TestConfigParser01(unittest.TestCase):
 
 
     def test_logging_global_params(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["logging_global"]
         have = parser.logging_global_params
@@ -101,7 +104,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_logging(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = {
             "sources": self.results["logging_global"]["sources"],
@@ -111,7 +114,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_tacacs_servers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["tacacs_servers"]
         have = parser.tacacs_servers
@@ -119,7 +122,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_radius_servers(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["radius_servers"]
         have = parser.radius_servers
@@ -127,7 +130,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_tacacs_groups(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["tacacs_groups"]
         have = parser.tacacs_groups
@@ -135,7 +138,7 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_radius_groups(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         want = self.results["radius_groups"]
         have = parser.radius_groups
@@ -143,19 +146,20 @@ class TestConfigParser01(unittest.TestCase):
         self.assertEqual(want, have)
 
     def test_aaa_login_methods(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         have = parser.aaa_login_methods
         jprint(have)
 
+
     def test_aaa_authorization_exec_methods(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         have = parser.aaa_authorization_exec_methods
         jprint(have)
 
     def test_isis(self):
-        parser = ConfigParser(config=self.config_path, device_type="ios")
+        parser = ConfigParser(config=self.config_path, device_type="ios", verbosity=VERBOSITY)
         parser.minimal_results = True
         have = parser.isis
         jprint(have)