+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## Version 0.2.18
+**Release Date:** 10-03-2021
+
+### Minor Changes
+Changed structure of `CiscoIosInterface.standby`, along with some new keys
+
+### Deprecation Warnings
+`CiscoIosInterface.ip_addresses` deprecated by `CiscoIosInterface.ipv4_addresses`
+
+`CiscoIosInterface.ip_unnumbered` deprecated by `CiscoIosInterface.ipv4_unnumbered`
+
 ## Version 0.2.16
 **Release Date:** 02-12-2020
 
@@ -23,7 +34,7 @@ Returns list of VLAN groups, such as
 ]
 ```
 #### CiscoIosInterface.dhcp_snooping
-*Note: This is just a quick one for specific use case. Currently only checks if interface is trusted (containing `ip dhcp snooping trus` child)*
+*Note: This is just a quick one for specific use case. Currently only checks if interface is trusted (containing `ip dhcp snooping trust` child)*
 
 In the future will return all info regarding DHCP Snooping on interface level. Currently only returns dict
 ```python